Moves SetDict into shared/utils. data_types as a lookup. Update validations after update of uijson. Add validate call to update uijson test fix discovered bug
