+++ geoh5py/ui_json/parameters.py
@@ -22,9 +22,8 @@ from uuid import UUID
 
 from geoh5py import Workspace
 from geoh5py.groups import PropertyGroup
+from geoh5py.shared.utils import SetDict
 from geoh5py.ui_json.enforcers import EnforcerPool
-
-from . import SetDict
 
 Validation = Dict[str, Any]
